% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flood_warnings.R
\name{get_flood_warnings}
\alias{get_flood_warnings}
\title{Get the current flood warnings from EA}
\usage{
get_flood_warnings(
  min_severity = NULL,
  county = NULL,
  lat = NULL,
  long = NULL,
  dist = NULL
)
}
\arguments{
\item{min_severity}{}
}
\value{

}
\description{
Get the current flood warnings from EA
}
