% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measures.R
\name{get_measures}
\alias{get_measures}
\title{Get details of measures available from the EA API}
\usage{
get_measures(
  parameter_name = NULL,
  parameter = NULL,
  qualifier = NULL,
  station_reference = NULL,
  station = NULL
)
}
\arguments{
\item{station_reference}{}
}
\value{

}
\description{
Get details of measures available from the EA API
}
