% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_measure.R
\name{get_single_measure}
\alias{get_single_measure}
\title{Get details of single measure from the EA API}
\usage{
get_single_measure(x)
}
\arguments{
\item{x}{Environment Agency API measure identifier}
}
\value{
A list of measure information.
}
\description{
Get details of single measure from the EA API
}
\examples{
get_single_measure("1491TH-level-stage-i-15_min-mASD")
}
