% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_station.R
\name{get_single_station}
\alias{get_single_station}
\title{Get details of single measurement station from the EA}
\usage{
get_single_station(x)
}
\arguments{
\item{x}{Environment Agency API Station identifier}
}
\value{
A list of station information
}
\description{
Get details of single measurement station from the EA
}
\examples{
get_single_station("1491TH")
}
