% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flood_warnings.R
\name{plot_flood_warnings}
\alias{plot_flood_warnings}
\title{Plot flood warnings on interactive map}
\usage{
plot_flood_warnings(
  min_severity = 3,
  county = NULL,
  lat = NULL,
  long = NULL,
  dist = NULL
)
}
\arguments{
\item{min_severity}{Return only warnings which at least as severe as this
level.}

\item{county}{Return only warnings applying to flood areas whose county name
contains the submitted string. Can be a list of county names separated by
"," in which case alerts which mention any of the named counties will be
returned.}

\item{lat, long, dist}{Return only warnings applying to flood areas which are
within \code{dist} km of the given latitude/longitude (in WGS84
coordinates), this may be approximated by a bounding box.}
}
\value{
A tmap object. Interactive exploration of flood warnings.
}
\description{
Plot flood warnings on interactive map
}
