% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measure_readings.R
\name{plot_measure_readings}
\alias{plot_measure_readings}
\title{Plot readings for a given measure}
\usage{
plot_measure_readings(
  x,
  today = FALSE,
  start_date = NULL,
  end_date = NULL,
  since = NULL,
  limit = NULL,
  show_max = TRUE,
  show_high = TRUE,
  show_low = TRUE
)
}
\arguments{
\item{x}{Measure URI.}

\item{today}{Return all the readings taken today for each included measure.}

\item{start_date, end_date}{Return the readings taken on the specified range
of days for each included measure, up to the specified limit.}

\item{since}{Return the readings taken since the given date time
(not inclusive), up to the specified limit. Will accept a simple date value
such as 2016-09-07 which will be interpreted as 2016-09-07T:00:00:00Z.}

\item{limit}{Return a maximum of these items from the list. By default the
readings API endpoints have a limit of 500 items, this can be modified by
providing an explicit limit value up to a hard limit of 10000 items.}

\item{show_max}{Where available, show max on record as reference line on
chart.}

\item{show_high}{Where available, show typical high as reference line on
chart.}

\item{show_low}{Where available, show typical low as reference line on
chart.}
}
\value{
ggplot2 plot object.
}
\description{
Readings for each of the published measures are available for the recent
period (up to the last four weeks). Longer historic data records may be
available separately. Each reading comprises a reference to the measure
being read (identified by its URI), a date time stamp for when the reading
applies and a numeric value. The set of readings is updated every 15
minutes, but the individual measures may be updated less frequently than
this.
}
