% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stations.R
\name{plot_stations}
\alias{plot_stations}
\title{Plot flood monitoring stations on a map}
\usage{
plot_stations(..., interactive = TRUE)
}
\arguments{
\item{...}{Other arguments passed to get_stations()}

\item{interactive}{If TRUE return an interactive map, otherwise a static
image.}
}
\value{
Leaflet map if interactive = TRUE, otherwise an image.
}
\description{
Plot flood monitoring stations on a map
}
