% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons_as_sf.R
\name{polygons_as_sf}
\alias{polygons_as_sf}
\title{Download Geo JSON polygons and return as single simple features (sf)}
\usage{
polygons_as_sf(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Download Geo JSON polygons and return as single simple features (sf)
}
