% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flood_areas.R
\name{get_flood_areas}
\alias{get_flood_areas}
\title{Get the flood areas from EA}
\usage{
get_flood_areas(lat = NULL, long = NULL, dist = NULL, limit = NULL)
}
\arguments{
\item{lat, long, dist}{Return only warnings applying to flood areas which are
within \code{dist} km of the given latitude/longitude (in WGS84
coordinates), this may be approximated by a bounding box.}

\item{limit}{The default limit on number of returned items is 500, this can
be adjusted by giving an explicit value, there is no hard upper limit.}
}
\value{
A dataframe of Environment Agency flood areas for the given criteria
}
\description{
Get the flood areas from EA
}
