% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flood_areas.R
\name{get_flood_areas}
\alias{get_flood_areas}
\title{Get the flood areas from EA}
\usage{
get_flood_areas(lat = NULL, long = NULL, dist = NULL, limit = NULL)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get the flood areas from EA
}
