% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_readings.R
\name{get_readings}
\alias{get_readings}
\title{Get readings from measurement stations via the EA API}
\usage{
get_readings(
  measure = NULL,
  station = NULL,
  latest = FALSE,
  today = FALSE,
  start_date = NULL,
  end_date = NULL,
  since = NULL,
  limit = NULL,
  sorted = FALSE
)
}
\arguments{
\item{limit}{}
}
\value{

}
\description{
Get readings from measurement stations via the EA API
}
