% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_measure.R
\name{get_single_measure}
\alias{get_single_measure}
\title{Get details of single measure from the EA API}
\usage{
get_single_measure(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get details of single measure from the EA API
}
