% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{Get details of measurement stations from the EA}
\usage{
get_stations(
  parameter_name = NULL,
  parameter = NULL,
  qualifier = NULL,
  label = NULL,
  town = NULL,
  catchment_name = NULL,
  river_name = NULL,
  station_reference = NULL,
  rloi_id = NULL,
  search = NULL,
  lat = NULL,
  long = NULL,
  dist = NULL,
  type = NULL,
  status = NULL
)
}
\arguments{
\item{status}{}
}
\value{

}
\description{
Get details of measurement stations from the EA
}
