% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flood_areas.R
\name{plot_flood_areas}
\alias{plot_flood_areas}
\title{Plot flood areas on interactive map}
\usage{
plot_flood_areas(lat = NULL, long = NULL, dist = NULL, limit = NULL)
}
\arguments{
\item{lat, long, dist}{Return only warnings applying to flood areas which are
within \code{dist} km of the given latitude/longitude (in WGS84
coordinates), this may be approximated by a bounding box.}

\item{limit}{}
}
\value{
A tmap object. Interactive exploration of flood warnings.
}
\description{
Plot flood areas on interactive map
}
