% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flood_warnings.R
\name{plot_flood_warnings}
\alias{plot_flood_warnings}
\title{Plot flood warnings on interactive map}
\usage{
plot_flood_warnings(
  min_severity = 3,
  county = NULL,
  lat = NULL,
  long = NULL,
  dist = NULL
)
}
\arguments{
\item{dist}{}
}
\value{

}
\description{
Plot flood warnings on interactive map
}
