% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measure_readings.R
\name{plot_measure_readings}
\alias{plot_measure_readings}
\title{Plot readings for a given measure}
\usage{
plot_measure_readings(x)
}
\arguments{
\item{x}{Measure URI.}
}
\value{
ggplot2 plot object.
}
\description{
Readings for each of the published measures are available for the recent
period (up to the last four weeks). Longer historic data records may be
available separately. Each reading comprises a reference to the measure
being read (identified by its URI), a date time stamp for when the reading
applies and a numeric value. The set of readings is updated every 15
minutes, but the individual measures may be updated less frequently than
this.
}
